requirements.txt
streamlit
gspread
oauth2client
pandas
gTTS
requests
fpdf
google-generativeai